rebase blender artists
