+++ src/compas_blender/__init__.py
@@ -47,4 +47,5 @@ __version__ = '1.8.1'
 __all__ = [name for name in dir() if not name.startswith('_')]
 __all_plugins__ = [
     # 'compas_blender.geometry.booleans',
+    'compas_blender.artists',
 ]
