+++ src/compas_blender/artists/networkartist.py
@@ -1,21 +1,14 @@
-# from __future__ import annotations
-
 from functools import partial
 
 import compas_blender
 
-from compas_blender.artists._artist import BaseArtist
 from compas.utilities import color_to_colordict
+from .artist import BlenderArtist
 
 colordict = partial(color_to_colordict, colorformat='rgb', normalize=True)
 
 
-__all__ = [
-    'NetworkArtist',
-]
-
-
-class NetworkArtist(BaseArtist):
+class NetworkArtist(BlenderArtist):
     """Artist for COMPAS network objects.
 
     Parameters
